+++ android/src/main/kotlin/com/eopeter/fluttermapboxnavigation/FlutterMapboxNavigationPlugin.kt
@@ -1,17 +1,15 @@
-package com.eopeter.flutter_mapbox_navigation
+package com.eopeter.fluttermapboxnavigation
 
 import android.Manifest
 import android.app.Activity
 import android.content.Context
 import android.content.pm.PackageManager
 import android.os.Build
-import com.eopeter.flutter_mapbox_navigation.activity.NavigationLauncher
-import com.eopeter.flutter_mapbox_navigation.factory.EmbeddedNavigationViewFactory
-import com.eopeter.flutter_mapbox_navigation.models.Waypoint
-
+import com.eopeter.fluttermapboxnavigation.activity.NavigationLauncher
+import com.eopeter.fluttermapboxnavigation.factory.EmbeddedNavigationViewFactory
+import com.eopeter.fluttermapboxnavigation.models.Waypoint
 import com.mapbox.api.directions.v5.DirectionsCriteria
 import com.mapbox.api.directions.v5.models.DirectionsRoute
-
 import io.flutter.embedding.engine.plugins.FlutterPlugin
 import io.flutter.embedding.engine.plugins.activity.ActivityAware
 import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding
@@ -22,11 +20,9 @@ import io.flutter.plugin.common.MethodChannel
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler
 import io.flutter.plugin.common.MethodChannel.Result
 import io.flutter.plugin.platform.PlatformViewRegistry
-import java.util.*
-import android.util.Log
 
 /** FlutterMapboxNavigationPlugin */
-public class FlutterMapboxNavigationPlugin : FlutterPlugin, MethodCallHandler,
+class FlutterMapboxNavigationPlugin : FlutterPlugin, MethodCallHandler,
     EventChannel.StreamHandler, ActivityAware {
 
     private lateinit var channel: MethodChannel
@@ -43,7 +39,9 @@ public class FlutterMapboxNavigationPlugin : FlutterPlugin, MethodCallHandler,
         progressEventChannel.setStreamHandler(this)
 
         platformViewRegistry = binding.platformViewRegistry
-        binaryMessenger = messenger;
+        binaryMessenger = messenger
+
+
     }
 
     companion object {
@@ -57,14 +55,18 @@ public class FlutterMapboxNavigationPlugin : FlutterPlugin, MethodCallHandler,
         val wayPoints: MutableList<Waypoint> = mutableListOf()
 
         var showAlternateRoutes: Boolean = true
-        val allowsClickToSetDestination: Boolean = false
+        var longPressDestinationEnabled: Boolean = true
         var allowsUTurnsAtWayPoints: Boolean = false
+        var enableOnMapTapCallback: Boolean = false
         var navigationMode = DirectionsCriteria.PROFILE_DRIVING_TRAFFIC
         var simulateRoute = false
+        var enableFreeDriveMode = false
         var mapStyleUrlDay: String? = null
         var mapStyleUrlNight: String? = null
         var navigationLanguage = "en"
         var navigationVoiceUnits = DirectionsCriteria.IMPERIAL
+        var voiceInstructionsEnabled = true
+        var bannerInstructionsEnabled = true
         var zoom = 15.0
         var bearing = 0.0
         var tilt = 0.0
@@ -82,12 +84,17 @@ public class FlutterMapboxNavigationPlugin : FlutterPlugin, MethodCallHandler,
                 result.success("Android ${Build.VERSION.RELEASE}")
             }
             "getDistanceRemaining" -> {
-                result.success(distanceRemaining);
+                result.success(distanceRemaining)
             }
             "getDurationRemaining" -> {
-                result.success(durationRemaining);
+                result.success(durationRemaining)
+            }
+            "startFreeDrive" -> {
+                enableFreeDriveMode = true
+                checkPermissionAndBeginNavigation(call)
             }
             "startNavigation" -> {
+                enableFreeDriveMode = false
                 checkPermissionAndBeginNavigation(call)
             }
             "addWayPoints" -> {
@@ -105,7 +112,7 @@ public class FlutterMapboxNavigationPlugin : FlutterPlugin, MethodCallHandler,
 
     private fun downloadRegionForOfflineRouting(
         call: MethodCall,
-        result: MethodChannel.Result
+        result: Result
     ) {
         result.error("TODO", "Not Implemented in Android", "will implement soon")
     }
@@ -121,7 +128,7 @@ public class FlutterMapboxNavigationPlugin : FlutterPlugin, MethodCallHandler,
                 "walking" -> navigationMode = DirectionsCriteria.PROFILE_WALKING
                 "cycling" -> navigationMode = DirectionsCriteria.PROFILE_CYCLING
                 "driving" -> navigationMode = DirectionsCriteria.PROFILE_DRIVING
-            };
+            }
         }
 
         val alternateRoutes = arguments?.get("alternatives") as? Boolean
@@ -139,23 +146,50 @@ public class FlutterMapboxNavigationPlugin : FlutterPlugin, MethodCallHandler,
             allowsUTurnsAtWayPoints = allowsUTurns
         }
 
+        val onMapTap = arguments?.get("enableOnMapTapCallback") as? Boolean
+        if (onMapTap != null) {
+            enableOnMapTapCallback = onMapTap
+        }
+
         val language = arguments?.get("language") as? String
-        if (language != null)
+        if (language != null) {
             navigationLanguage = language
+        }
+
+        val voiceEnabled = arguments?.get("voiceInstructionsEnabled") as? Boolean
+        if (voiceEnabled != null) {
+            voiceInstructionsEnabled = voiceEnabled
+        }
+
+        val bannerEnabled = arguments?.get("bannerInstructionsEnabled") as? Boolean
+        if (bannerEnabled != null) {
+            bannerInstructionsEnabled = bannerEnabled
+        }
 
         val units = arguments?.get("units") as? String
 
         if (units != null) {
-            if (units == "imperial")
+            if (units == "imperial") {
                 navigationVoiceUnits = DirectionsCriteria.IMPERIAL
-            else if (units == "metric")
+            } else if (units == "metric") {
                 navigationVoiceUnits = DirectionsCriteria.METRIC
+            }
         }
 
         mapStyleUrlDay = arguments?.get("mapStyleUrlDay") as? String
         mapStyleUrlNight = arguments?.get("mapStyleUrlNight") as? String
 
+        val longPress = arguments?.get("longPressDestinationEnabled") as? Boolean
+        if (longPress != null) {
+            longPressDestinationEnabled = longPress
+        }
+
         wayPoints.clear()
+
+        if (enableFreeDriveMode) {
+            checkPermissionAndBeginNavigation(wayPoints)
+            return
+        }
 
         val points = arguments?.get("wayPoints") as HashMap<Int, Any>
         for (item in points) {
@@ -166,9 +200,7 @@ public class FlutterMapboxNavigationPlugin : FlutterPlugin, MethodCallHandler,
             val isSilent = point["IsSilent"] as Boolean
             wayPoints.add(Waypoint(name, longitude, latitude, isSilent))
         }
-
         checkPermissionAndBeginNavigation(wayPoints)
-
     }
 
     private fun checkPermissionAndBeginNavigation(wayPoints: List<Waypoint>) {
@@ -189,12 +221,12 @@ public class FlutterMapboxNavigationPlugin : FlutterPlugin, MethodCallHandler,
     }
 
     private fun beginNavigation(wayPoints: List<Waypoint>) {
-        NavigationLauncher.startNavigation(currentActivity, wayPoints);
+        NavigationLauncher.startNavigation(currentActivity, wayPoints)
     }
 
     private fun addWayPointsToNavigation(
         call: MethodCall,
-        result: MethodChannel.Result
+        result: Result
     ) {
         val arguments = call.arguments as? Map<String, Any>
         val points = arguments?.get("wayPoints") as HashMap<Int, Any>
@@ -207,23 +239,21 @@ public class FlutterMapboxNavigationPlugin : FlutterPlugin, MethodCallHandler,
             val isSilent = point["IsSilent"] as Boolean
             wayPoints.add(Waypoint(name, latitude, longitude, isSilent))
         }
-        NavigationLauncher.addWayPoints(currentActivity, wayPoints);
+        NavigationLauncher.addWayPoints(currentActivity, wayPoints)
     }
 
     override fun onListen(args: Any?, events: EventChannel.EventSink?) {
-        eventSink = events;
+        eventSink = events
     }
 
     override fun onCancel(args: Any?) {
-        eventSink = null;
+        eventSink = null
     }
 
     override fun onDetachedFromEngine(binding: FlutterPlugin.FlutterPluginBinding) {
-
         currentActivity = null
         channel.setMethodCallHandler(null)
         progressEventChannel.setStreamHandler(null)
-
     }
 
     override fun onDetachedFromActivity() {
@@ -232,15 +262,12 @@ public class FlutterMapboxNavigationPlugin : FlutterPlugin, MethodCallHandler,
     }
 
     override fun onReattachedToActivityForConfigChanges(binding: ActivityPluginBinding) {
-
         currentActivity = binding.activity
-
     }
 
     override fun onAttachedToActivity(binding: ActivityPluginBinding) {
         currentActivity = binding.activity
         currentContext = binding.activity.applicationContext
-
         if (platformViewRegistry != null && binaryMessenger != null && currentActivity != null) {
             platformViewRegistry?.registerViewFactory(
                 viewId,
@@ -250,7 +277,7 @@ public class FlutterMapboxNavigationPlugin : FlutterPlugin, MethodCallHandler,
     }
 
     override fun onDetachedFromActivityForConfigChanges() {
-        //To change body of created functions use File | Settings | File Templates.
+        // To change body of created functions use File | Settings | File Templates.
     }
 
     fun onRequestPermissionsResult(
@@ -260,7 +287,6 @@ public class FlutterMapboxNavigationPlugin : FlutterPlugin, MethodCallHandler,
     ) {
         when (requestCode) {
             367 -> {
-
                 for (permission in permissions) {
                     if (permission == Manifest.permission.ACCESS_FINE_LOCATION) {
                         val haspermission = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
@@ -269,7 +295,7 @@ public class FlutterMapboxNavigationPlugin : FlutterPlugin, MethodCallHandler,
                             TODO("VERSION.SDK_INT < M")
                         }
                         if (haspermission == PackageManager.PERMISSION_GRANTED) {
-                            if (wayPoints.count() > 0)
+                            if (wayPoints.isNotEmpty())
                                 beginNavigation(wayPoints)
                         }
                         // Not all permissions granted. Show some message and return.
@@ -280,10 +306,8 @@ public class FlutterMapboxNavigationPlugin : FlutterPlugin, MethodCallHandler,
                 // All permissions are granted. Do the work accordingly.
             }
         }
-        //super.onRequestPermissionsResult(requestCode, permissions, grantResults)
-
+        // super.onRequestPermissionsResult(requestCode, permissions, grantResults)
     }
-
 }
 
 private const val MAPBOX_ACCESS_TOKEN_PLACEHOLDER = "YOUR_MAPBOX_ACCESS_TOKEN_GOES_HERE"