+++ android/src/main/kotlin/com/eopeter/fluttermapboxnavigation/models/MapBoxEvents.kt
@@ -1,4 +1,4 @@
-package com.eopeter.flutter_mapbox_navigation.models
+package com.eopeter.fluttermapboxnavigation.models
 
 enum class MapBoxEvents(val value: String) {
     MAP_READY("map_ready"),
@@ -19,4 +19,5 @@ enum class MapBoxEvents(val value: String) {
     ON_ARRIVAL("on_arrival"),
     FAILED_TO_REROUTE("failed_to_reroute"),
     REROUTE_ALONG("reroute_along"),
+    ON_MAP_TAP("on_map_tap")
 }