+++ example/lib/app.dart
@@ -1,4 +1,5 @@
 import 'dart:async';
+
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_mapbox_navigation/flutter_mapbox_navigation.dart';
@@ -7,7 +8,7 @@ class SampleNavigationApp extends StatefulWidget {
   const SampleNavigationApp({super.key});
 
   @override
-  _SampleNavigationAppState createState() => _SampleNavigationAppState();
+  State<SampleNavigationApp> createState() => _SampleNavigationAppState();
 }
 
 class _SampleNavigationAppState extends State<SampleNavigationApp> {
@@ -56,12 +57,19 @@ class _SampleNavigationAppState extends State<SampleNavigationApp> {
   MapBoxNavigationViewController? _controller;
   bool _routeBuilt = false;
   bool _isNavigating = false;
+  bool _inFreeDrive = false;
   late MapBoxOptions _navigationOption;
 
   @override
   void initState() {
     super.initState();
     initialize();
+  }
+
+  @override
+  void dispose() {
+    _controller?.dispose();
+    super.dispose();
   }
 
   // Platform messages are asynchronous, so we initialize in an async method.
@@ -72,11 +80,11 @@ class _SampleNavigationAppState extends State<SampleNavigationApp> {
     if (!mounted) return;
 
     _navigationOption = MapBoxNavigation.instance.getDefaultOptions();
-    _navigationOption.simulateRoute = false;
+    _navigationOption.simulateRoute = true;
+    _navigationOption.language = "en";
     //_navigationOption.initialLatitude = 36.1175275;
     //_navigationOption.initialLongitude = -115.1839524;
     MapBoxNavigation.instance.registerRouteEventListener(_onEmbeddedRouteEvent);
-    MapBoxNavigation.instance.setDefaultOptions(_navigationOption);
 
     String? platformVersion;
     // Platform messages may fail, so we use a try/catch PlatformException.
@@ -86,9 +94,9 @@ class _SampleNavigationAppState extends State<SampleNavigationApp> {
       platformVersion = 'Failed to get platform version.';
     }
 
-    // setState(() {
-    //   _platformVersion = platformVersion;
-    // });
+    setState(() {
+      _platformVersion = platformVersion;
+    });
   }
 
   @override
@@ -98,43 +106,227 @@ class _SampleNavigationAppState extends State<SampleNavigationApp> {
         appBar: AppBar(
           title: const Text('Plugin example app'),
         ),
-        body: SingleChildScrollView(
-          child: Column(
-            children: [
-              const SizedBox(
-                height: 10,
-              ),
-              Text('Running on: $_platformVersion\n'),
-              Container(
-                color: Colors.grey,
-                width: double.infinity,
-                child: const Padding(
-                  padding: EdgeInsets.all(10),
-                  child: Text(
-                    "Full Screen Navigation",
-                    style: TextStyle(color: Colors.white),
-                    textAlign: TextAlign.center,
-                  ),
+        body: Center(
+          child: Column(children: <Widget>[
+            Expanded(
+              child: SingleChildScrollView(
+                child: Column(
+                  children: [
+                    const SizedBox(
+                      height: 10,
+                    ),
+                    Text('Running on: $_platformVersion\n'),
+                    Container(
+                      color: Colors.grey,
+                      width: double.infinity,
+                      child: const Padding(
+                        padding: EdgeInsets.all(10),
+                        child: (Text(
+                          "Full Screen Navigation",
+                          style: TextStyle(color: Colors.white),
+                          textAlign: TextAlign.center,
+                        )),
+                      ),
+                    ),
+                    Row(
+                      mainAxisAlignment: MainAxisAlignment.center,
+                      children: [
+                        ElevatedButton(
+                          child: const Text("Start A to B"),
+                          onPressed: () async {
+                            var wayPoints = <WayPoint>[];
+                            wayPoints.add(_home);
+                            wayPoints.add(_store);
+                            var opt = MapBoxOptions.from(_navigationOption);
+                            opt.simulateRoute = true;
+                            opt.voiceInstructionsEnabled = true;
+                            opt.bannerInstructionsEnabled = true;
+                            opt.units = VoiceUnits.metric;
+                            opt.language = "de-DE";
+                            await MapBoxNavigation.instance
+                                .startNavigation(wayPoints: wayPoints, options: opt);
+                          },
+                        ),
+                        const SizedBox(
+                          width: 10,
+                        ),
+                        ElevatedButton(
+                          child: const Text("Start Multi Stop"),
+                          onPressed: () async {
+                            _isMultipleStop = true;
+                            var wayPoints = <WayPoint>[];
+                            wayPoints.add(_origin);
+                            wayPoints.add(_stop1);
+                            wayPoints.add(_stop2);
+                            wayPoints.add(_stop3);
+                            wayPoints.add(_destination);
+
+                            MapBoxNavigation.instance.startNavigation(
+                                wayPoints: wayPoints,
+                                options: MapBoxOptions(
+                                    mode: MapBoxNavigationMode.driving,
+                                    simulateRoute: true,
+                                    language: "en",
+                                    allowsUTurnAtWayPoints: true,
+                                    units: VoiceUnits.metric));
+                            //after 10 seconds add a new stop
+                            await Future.delayed(const Duration(seconds: 10));
+                            var stop = WayPoint(
+                                name: "Gas Station",
+                                latitude: 38.911176544398,
+                                longitude: -77.04014366543564,
+                                isSilent: false);
+                            MapBoxNavigation.instance
+                                .addWayPoints(wayPoints: [stop]);
+                          },
+                        ),
+                        const SizedBox(
+                          width: 10,
+                        ),
+                        ElevatedButton(
+                          child: const Text("Free Drive"),
+                          onPressed: () async {
+                            await MapBoxNavigation.instance.startFreeDrive();
+                          },
+                        ),
+                      ],
+                    ),
+                    Container(
+                      color: Colors.grey,
+                      width: double.infinity,
+                      child: const Padding(
+                        padding: EdgeInsets.all(10),
+                        child: (Text(
+                          "Embedded Navigation",
+                          style: TextStyle(color: Colors.white),
+                          textAlign: TextAlign.center,
+                        )),
+                      ),
+                    ),
+                    Row(
+                      mainAxisAlignment: MainAxisAlignment.center,
+                      children: [
+                        ElevatedButton(
+                          onPressed: _isNavigating
+                              ? null
+                              : () {
+                                  if (_routeBuilt) {
+                                    _controller?.clearRoute();
+                                  } else {
+                                    var wayPoints = <WayPoint>[];
+                                    wayPoints.add(_home);
+                                    wayPoints.add(_store);
+                                    _isMultipleStop = wayPoints.length > 2;
+                                    _controller?.buildRoute(
+                                        wayPoints: wayPoints,
+                                        options: _navigationOption);
+                                  }
+                                },
+                          child: Text(_routeBuilt && !_isNavigating
+                              ? "Clear Route"
+                              : "Build Route"),
+                        ),
+                        const SizedBox(
+                          width: 10,
+                        ),
+                        ElevatedButton(
+                          onPressed: _routeBuilt && !_isNavigating
+                              ? () {
+                                  _controller?.startNavigation();
+                                }
+                              : null,
+                          child: const Text('Start '),
+                        ),
+                        const SizedBox(
+                          width: 10,
+                        ),
+                        ElevatedButton(
+                          onPressed: _isNavigating
+                              ? () {
+                                  _controller?.finishNavigation();
+                                }
+                              : null,
+                          child: const Text('Cancel '),
+                        )
+                      ],
+                    ),
+                    ElevatedButton(
+                      onPressed: _inFreeDrive
+                          ? null
+                          : () async {
+                              _inFreeDrive =
+                                  await _controller?.startFreeDrive() ?? false;
+                            },
+                      child: const Text("Free Drive "),
+                    ),
+                    const Center(
+                      child: Padding(
+                        padding: EdgeInsets.all(10),
+                        child: Text(
+                          "Long-Press Embedded Map to Set Destination",
+                          textAlign: TextAlign.center,
+                        ),
+                      ),
+                    ),
+                    Container(
+                      color: Colors.grey,
+                      width: double.infinity,
+                      child: Padding(
+                        padding: const EdgeInsets.all(10),
+                        child: (Text(
+                          _instruction == null
+                              ? "Banner Instruction Here"
+                              : _instruction!,
+                          style: const TextStyle(color: Colors.white),
+                          textAlign: TextAlign.center,
+                        )),
+                      ),
+                    ),
+                    Padding(
+                      padding: const EdgeInsets.only(
+                          left: 20.0, right: 20, top: 20, bottom: 10),
+                      child: Column(
+                        mainAxisAlignment: MainAxisAlignment.center,
+                        children: <Widget>[
+                          Row(
+                            children: <Widget>[
+                              const Text("Duration Remaining: "),
+                              Text(_durationRemaining != null
+                                  ? "${(_durationRemaining! / 60).toStringAsFixed(0)} minutes"
+                                  : "---")
+                            ],
+                          ),
+                          Row(
+                            children: <Widget>[
+                              const Text("Distance Remaining: "),
+                              Text(_distanceRemaining != null
+                                  ? "${(_distanceRemaining! * 0.000621371).toStringAsFixed(1)} miles"
+                                  : "---")
+                            ],
+                          ),
+                        ],
+                      ),
+                    ),
+                    const Divider()
+                  ],
                 ),
               ),
-              Row(
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: [
-                  ElevatedButton(
-                    child: const Text("Start A to B"),
-                    onPressed: () async {
-                      var wayPoints = <WayPoint>[];
-                      wayPoints.add(_home);
-                      wayPoints.add(_store);
-
-                      await MapBoxNavigation.instance
-                          .startNavigation(wayPoints: wayPoints);
-                    },
-                  ),
-                ],
+            ),
+            SizedBox(
+              height: 300,
+              child: Container(
+                color: Colors.grey,
+                child: MapBoxNavigationView(
+                    options: _navigationOption,
+                    onRouteEvent: _onEmbeddedRouteEvent,
+                    onCreated:
+                        (MapBoxNavigationViewController controller) async {
+                      _controller = controller;
+                      controller.initialize();
+                    }),
               ),
-            ],
-          ),
+            )
+          ]),
         ),
       ),
     );
@@ -183,7 +375,6 @@ class _SampleNavigationAppState extends State<SampleNavigationApp> {
       default:
         break;
     }
-
     setState(() {});
   }
 }