+++ lib/src/models/options.dart
@@ -1,34 +1,111 @@
-import 'package:flutter/widgets.dart';
+// ignore_for_file: public_member_api_docs
 
-import 'navmode.dart';
-import 'voice_units.dart';
+import 'package:flutter/widgets.dart';
+import 'package:flutter_mapbox_navigation/src/models/navmode.dart';
+import 'package:flutter_mapbox_navigation/src/models/voice_units.dart';
 
 /// Configuration options for the MapBoxNavigation.
 ///
 /// When used to change configuration, null values will be interpreted as
-/// "do not change this configuration option".
+/// 'do not change this configuration option'.
 ///
 class MapBoxOptions {
+
+  MapBoxOptions({
+    this.initialLatitude,
+    this.initialLongitude,
+    this.language,
+    this.zoom,
+    this.bearing,
+    this.tilt,
+    this.alternatives,
+    this.mode,
+    this.units,
+    this.allowsUTurnAtWayPoints,
+    this.enableRefresh,
+    this.voiceInstructionsEnabled,
+    this.bannerInstructionsEnabled,
+    this.longPressDestinationEnabled,
+    this.simulateRoute,
+    this.isOptimized,
+    this.mapStyleUrlDay,
+    this.mapStyleUrlNight,
+    this.padding,
+    this.animateBuildRoute,
+    this.showReportFeedbackButton = true,
+    this.showEndOfRouteFeedback = true,
+    this.enableOnMapTapCallback = false,
+  });
+
+  MapBoxOptions.from(MapBoxOptions option) {
+    initialLatitude = option.initialLatitude;
+    initialLongitude = option.initialLongitude;
+    language = option.language;
+    zoom = option.zoom;
+    bearing = option.bearing;
+    tilt = option.tilt;
+    alternatives = option.alternatives;
+    mode = option.mode;
+    units = option.units;
+    allowsUTurnAtWayPoints = option.allowsUTurnAtWayPoints;
+    enableRefresh = option.enableRefresh;
+    voiceInstructionsEnabled = option.voiceInstructionsEnabled;
+    bannerInstructionsEnabled = option.bannerInstructionsEnabled;
+    longPressDestinationEnabled = option.longPressDestinationEnabled;
+    simulateRoute = option.simulateRoute;
+    isOptimized = option.isOptimized;
+    mapStyleUrlDay = option.mapStyleUrlDay;
+    mapStyleUrlNight = option.mapStyleUrlNight;
+    padding = option.padding;
+    animateBuildRoute = option.animateBuildRoute;
+    showReportFeedbackButton = option.showReportFeedbackButton;
+    showEndOfRouteFeedback = option.showEndOfRouteFeedback;
+  }
+
   /// The initial Latitude of the Map View
   double? initialLatitude;
 
   /// The initial Longitude of the Map View
   double? initialLongitude;
 
-  /// 2-letter ISO 639-1 code for language. This property affects the sentence contained within the RouteStep.instructions property, but it does not affect any road names contained in that property or other properties such as RouteStep.name. Defaults to "en" if an unsupported language is specified. The languages in this link are supported: https://docs.mapbox.com/android/navigation/overview/localization/ or https://docs.mapbox.com/ios/api/navigation/0.14.1/localization-and-internationalization.html
+  /// 2-letter ISO 639-1 code for language and optionally append the ISO-3166
+  /// country code for specific dialect like pt-BR for Brazilian Portuguese
+  /// This property affects the sentence
+  /// contained within the RouteStep.instructions property, but it does not
+  /// affect any road names contained in that property or other properties
+  /// such as RouteStep.name. Defaults to 'en' if an unsupported language
+  /// is specified. The languages in this link are supported:
+  ///  https://docs.mapbox.com/android/navigation/overview/localization/ or
+  /// https://docs.mapbox.com/ios/api/navigation/0.14.1/localization-and-internationalization.html
   String? language;
 
-  /// Zoom controls the scale of the map and consumes any value between 0 and 22. At zoom level 0, the viewport shows continents and other world features. A middle value of 11 will show city level details, and at a higher zoom level, the map will begin to show buildings and points of interest.
+  /// Zoom controls the scale of the map and consumes any value between 0
+  /// and 22. At zoom level 0, the viewport shows continents and other world
+  /// features. A middle value of 11 will show city level details, and at a
+  ///  higher zoom level, the map will begin to show buildings and
+  /// points of interest.
   double? zoom;
 
-  /// Bearing is the direction that the camera is pointing in and measured in degrees clockwise from north.
+  /// Bearing is the direction that the camera is pointing in and
+  /// measured in degrees clockwise from north.
   ///
-  /// The camera's default bearing is 0 degrees (i.e. "true north") causing the map compass to hide until the camera bearing becomes a non-zero value. The mapbox_uiCompass boolean XML attribute allows adjustment of the compass' visibility. Bearing levels use six decimal point precision, which enables you to restrict/set/lock a map's bearing with extreme precision. Besides programmatically adjusting the camera bearing, the user can place two fingertips on the map and rotate their fingers.
+  /// The camera's default bearing is 0 degrees (i.e. 'true north') causing the
+  /// map compass to hide until the camera bearing becomes a non-zero value.
+  /// The mapbox_uiCompass boolean XML attribute allows adjustment of
+  /// the compass' visibility. Bearing levels use six decimal point precision,
+  /// which enables you to restrict/set/lock a map's bearing with
+  /// extreme precision. Besides programmatically adjusting the camera bearing,
+  ///  the user can place two fingertips on the map and rotate their fingers.
   double? bearing;
 
-  /// Tilt is the camera's angle from the nadir (directly facing the Earth) and uses unit degrees. The camera's minimum (default) tilt is 0 degrees, and the maximum tilt is 60. Tilt levels use six decimal point of precision, which enables you to restrict/set/lock a map's bearing with extreme precision.
+  /// Tilt is the camera's angle from the nadir (directly facing the Earth)
+  /// and uses unit degrees. The camera's minimum (default) tilt is 0 degrees,
+  /// and the maximum tilt is 60. Tilt levels use six decimal point of
+  /// precision, which enables you to restrict/set/lock a map's bearing
+  /// with extreme precision.
   ///
-  /// The map camera tilt can also adjust by placing two fingertips on the map and moving both fingers up and down in parallel at the same time or
+  /// The map camera tilt can also adjust by placing two fingertips on the map
+  /// and moving both fingers up and down in parallel at the same time or
   double? tilt;
 
   ///
@@ -42,17 +119,22 @@ class MapBoxOptions {
   /// The unit of measure said in voice instructions
   VoiceUnits? units;
 
-  /// If the value of this property is true, a returned route may require an immediate U-turn at an intermediate waypoint. At an intermediate waypoint, if the value of this property is false, each returned route may continue straight ahead or turn to either side but may not U-turn. This property has no effect if only two waypoints are specified.
+  /// If the value of this property is true, a returned route may require an
+  /// immediate U-turn at an intermediate waypoint. At an intermediate waypoint,
+  ///  if the value of this property is false, each returned route may continue
+  /// straight ahead or turn to either side but may not U-turn. This property
+  /// has no effect if only two waypoints are specified.
   /// same as 'not continueStraight' on Android
   bool? allowsUTurnAtWayPoints;
 
-  final bool? enableRefresh;
+  bool? enableRefresh;
   // if true voice instruction is enabled
   bool? voiceInstructionsEnabled;
   //if true, banner instruction is shown and returned
   bool? bannerInstructionsEnabled;
 
-  /// if true will simulate the route as if you were driving. Always true on iOS Simulator
+  /// if true will simulate the route as if you were driving.
+  /// Always true on iOS Simulator
   bool? simulateRoute;
 
   /// The Url of the style the Navigation MapView should use during the day
@@ -61,7 +143,9 @@ class MapBoxOptions {
   /// The Url of the style the Navigation MapView should use at night
   String? mapStyleUrlNight;
 
-  /// if true, will reorder the routes to optimize navigation for time and shortest distance using the Travelling Salesman Algorithm. Always false for now
+  /// if true, will reorder the routes to optimize navigation for time and
+  /// shortest distance using the Travelling Salesman Algorithm.
+  /// Always false for now
   bool? isOptimized;
 
   /// Padding applied to the MapView when embedded
@@ -70,53 +154,33 @@ class MapBoxOptions {
   /// Should animate the building of the Route. Default is True
   bool? animateBuildRoute;
 
-  /// When the user long presses on a point on the map, set that as the destination
+  /// When the user long presses on a point on the map, set that
+  ///  as the destination
   bool? longPressDestinationEnabled;
 
-  /// Free-drive mode is a unique Mapbox Navigation SDK feature that allows drivers to navigate without a set destination. This mode is sometimes referred to as passive navigation.
-  /// No destination is required when set to true.
-  bool? enableFreeDriveMode;
+  /// In iOS this will show/hide the feedback button. Default to True.
+  bool? showReportFeedbackButton;
 
-  String? accessToken;
+  /// In iOS this will show/hide the end of route page when navigation is done. Default to True.
+  bool? showEndOfRouteFeedback;
 
-  MapBoxOptions(
-      {this.initialLatitude,
-      this.initialLongitude,
-      this.language,
-      this.zoom,
-      this.bearing,
-      this.tilt,
-      this.alternatives,
-      this.mode,
-      this.units,
-      this.allowsUTurnAtWayPoints,
-      this.enableRefresh,
-      this.voiceInstructionsEnabled,
-      this.bannerInstructionsEnabled,
-      this.longPressDestinationEnabled,
-      this.simulateRoute,
-      this.isOptimized,
-      this.mapStyleUrlDay,
-      this.mapStyleUrlNight,
-      this.enableFreeDriveMode,
-      this.padding,
-      this.animateBuildRoute,
-      this.accessToken});
+  /// Gives you the ability to receive back a waypoint corresponding
+  /// to where you tap on the map.
+  bool? enableOnMapTapCallback;
 
   Map<String, dynamic> toMap() {
-    final Map<String, dynamic> optionsMap = new Map<String, dynamic>();
+    final optionsMap = <String, dynamic>{};
     void addIfNonNull(String fieldName, dynamic value) {
       if (value != null) {
         optionsMap[fieldName] = value;
       }
     }
 
-    addIfNonNull("initialLatitude", initialLatitude);
-    addIfNonNull("initialLongitude", initialLongitude);
-    addIfNonNull("language", language);
-    addIfNonNull("animateBuildRoute", animateBuildRoute);
-    addIfNonNull("longPressDestinationEnabled", longPressDestinationEnabled);
-    addIfNonNull('accessToken', accessToken);
+    addIfNonNull('initialLatitude', initialLatitude);
+    addIfNonNull('initialLongitude', initialLongitude);
+    addIfNonNull('language', language);
+    addIfNonNull('animateBuildRoute', animateBuildRoute);
+    addIfNonNull('longPressDestinationEnabled', longPressDestinationEnabled);
 
     if (zoom != null) optionsMap['zoom'] = zoom;
     if (bearing != null) optionsMap['bearing'] = bearing;
@@ -137,8 +201,8 @@ class MapBoxOptions {
       optionsMap['enableRefresh'] = enableRefresh;
     }
 
-    addIfNonNull("voiceInstructionsEnabled", voiceInstructionsEnabled);
-    addIfNonNull("bannerInstructionsEnabled", bannerInstructionsEnabled);
+    addIfNonNull('voiceInstructionsEnabled', voiceInstructionsEnabled);
+    addIfNonNull('bannerInstructionsEnabled', bannerInstructionsEnabled);
 
     if (mapStyleUrlDay != null) {
       optionsMap['mapStyleUrlDay'] = mapStyleUrlDay;
@@ -158,14 +222,19 @@ class MapBoxOptions {
       padding?.right,
     ]);
 
+    addIfNonNull('showReportFeedbackButton', showReportFeedbackButton);
+    addIfNonNull('showEndOfRouteFeedback', showEndOfRouteFeedback);
+    addIfNonNull('enableOnMapTapCallback', enableOnMapTapCallback);
+
     return optionsMap;
   }
 
   Map<String, dynamic> updatesMap(MapBoxOptions newOptions) {
-    final Map<String, dynamic> prevOptionsMap = toMap();
+    final prevOptionsMap = toMap();
 
     return newOptions.toMap()
       ..removeWhere(
-          (String key, dynamic value) => prevOptionsMap[key] == value);
+        (String key, dynamic value) => prevOptionsMap[key] == value,
+      );
   }
 }